+++ src/main/java/com/publichealthnonprofit/programfunding/controller/service/FinancialGrantService.java
@@ -143,7 +143,7 @@ public class FinancialGrantService {
     private void setUpdatedFieldsInFinancialGrant(FinancialGrant financialGrant, FinancialGrantData financialGrantData) {
         FinancialGrantGrantingOrg updatedGrantingOrg = financialGrantData.getGrantingOrg();
         String updatedFinancialGrantName = financialGrantData.getFinancialGrantName();
-        Float updatedFinancialGrantAmount = financialGrantData.getFinancialGrantAmount();
+        Double updatedFinancialGrantAmount = financialGrantData.getFinancialGrantAmount();
         Date updatedFinancialGrantStartDate = financialGrantData.getFinancialGrantStartDate();
         Date updatedFinancialGrantEndDate = financialGrantData.getFinancialGrantEndDate();
         if (Objects.nonNull(updatedGrantingOrg)) {
